webapp/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
google-cloud-storage>=2.0
python-multipart>=0.0.9
